App/CreateRoomScreen - + extracted TreasuresListView to it's own file
